backend/src/main/java/com/sjiwon/anotherart/member/service/MemberInformationService.java
package com.sjiwon.anotherart.member.service;

import com.sjiwon.anotherart.member.domain.Member;
import com.sjiwon.anotherart.member.service.dto.response.MemberInformation;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional(readOnly = true)
@RequiredArgsConstructor
public class MemberInformationService {
    private final MemberFindService memberFindService;

    public MemberInformation getInformation(Long memberId) {
        Member member = memberFindService.findById(memberId);
        return new MemberInformation(member);
    }
}
